+++ plugin/src/net/groboclown/idea/p4ic/config/part/ClientNameDataPart.java
@@ -76,7 +76,7 @@ public class ClientNameDataPart extends DataPartAdapter {
     public Collection<ConfigProblem> getConfigProblems() {
         List<ConfigProblem> problems = new ArrayList<ConfigProblem>(additionalProblems);
         if (clientName == null) {
-            problems.add(new ConfigProblem(this, "error.config.no-client"));
+            problems.add(new ConfigProblem(this, false, "error.config.no-client"));
         }
         return problems;
     }